Create a reusable tab and set a dynamic component type
If there are several buttons with some action on them (e.g. onClick), instead of passing them all as props to the component, the button can be set as an attribute value in the parent component, where there is direct access to action function. This attribute is then passed as a prop.

To create a dynamic component type, an attribute with the type name as a string if built-in or in curly brackets as a custom component is created and passed as a prop.

To use the prop as a component type in the child component, either the attribute's name starts with a capital letter or (in the child component) a constant for the passed prop is created starting with a capital letter and then used as a component type.